+++ src/network/model/packet-metadata.cc
@@ -1248,8 +1248,8 @@ PacketMetadata::Serialize(uint8_t* buffer, uint32_t maxSize) const
         current = item.next;
     }
 
-    NS_ASSERT(static_cast<uint32_t>(buffer - start) == maxSize);
-    return 1;
+    NS_ASSERT(static_cast<uint32_t>(buffer - start) <= maxSize);
+    return static_cast<uint32_t>(buffer - start);
 }
 
 uint32_t
@@ -1257,7 +1257,7 @@ PacketMetadata::Deserialize(const uint8_t* buffer, uint32_t size)
 {
     NS_LOG_FUNCTION(this << &buffer << size);
     const uint8_t* start = buffer;
-    uint32_t desSize = size - 4;
+    uint32_t desSize = size;
 
     buffer = ReadFromRawU64(m_packetUid, start, buffer, size);
     desSize -= 8;
@@ -1310,7 +1310,7 @@ PacketMetadata::Deserialize(const uint8_t* buffer, uint32_t size)
         UpdateTail(tmp);
     }
     NS_ASSERT(desSize == 0);
-    return (desSize != 0) ? 0 : 1;
+    return (desSize != 0) ? 0 : size;
 }
 
 uint8_t*